requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
beautifulsoup4
mysql-connector-python